new coffee script / node.js branch
The python solution is okayish, but for JSON parsing as a very small
servic node.js is best

If you are not familiar with coffee script (a language that compiles to
javascript): http://js2.coffee/ there you can write js and view coffee
and the other way around. Its basicly improved javascript syntax.

This approach minimize the code (less code, less bugs, less to
maintain) in an modern functional javascript way